Update patch set 2
Patch Set 2:

( If you're not already familiar with the Zip task, you'll want to see the docs at https://docs.gradle.org/current/dsl/org.gradle.api.tasks.bundling.Zip.html and an example at https://android.googlesource.com/platform/frameworks/support/+/3cd0b2f87f02447c34ddae438dce1eca4d94aeab/buildSrc/src/main/kotlin/androidx/build/DiffAndDocs.kt#447 )



It is possible to upload to remote servers via Gradle, but uploading to Buganizer seems unusual enough to me that it seems reasonable to me for the upload to be part of a python script (and which also makes it easy to remember that Gradle truly only does "build") ; I'd be supportive of having the uploader and runner be in python. I imagine we would still want the zipping to be part of the Gradle build, though, partially to make it easier to parallelize it in the future if we want, and partially to reduce the distribution of languages in our build/release process.

I don't envision making buildOnServer depend on the new zip task any time soon (just having it as an easy option) ; you'll notice that buildOnServer currently doesn't depend on the (unzipped) diffs task either. The things that buildOnServer does depend on can mostly be found by searching in AndroidXPlugin.kt for the text "buildOnServerTask.dependsOn".

I expect it to be easier for the person that is breaking something to know how to fix it than the person that relies on the tool that is breaking. It's still possible for the former person to temporarily disable the code (and cc the latter person) if breaking it is intentional. This specific example is less significant because it's a small amount of code but I imagine that would make it less effort to move it, too.

Patch-set: 2
Reviewer: Gerrit User 1137668 <1137668@85c56323-6fa9-3386-8a01-6480fb634889>