Lazy loading of MediaRouteButton drawables
Animation drawables in MediaRouteButton doesn't need to be loaded if
the user do not want to cast.(especially on low-end devices)

So creates static drawables for initial display and loads
animation drawables when they're necessary.

Static drawables are also loaded asynchronously.

Also fixes a bug that the button displays "connected" just before
it begins connecting.

Bug: 30895587
Test: Ran Support v7 demos manually.
Change-Id: Id96bf0a680055996cd34df939ac9762c5e984b2b