Make IMediaController2.onPlaylistChanged use ParcelImplListSlice
This CL changes the argument of IMediaController2.onPlaylistChanged from
List<ParcelImpl> to ParcelImplListSlice. By this change, MediaSession2
can send a long playlist to MediaController2s when it is changed.

Note: Below API 21, MediaControllerCompat which is connected to
MediaSession2 will get only part of the playlist from the beginning
if the playlist is too large.

Bug: 112455503
Test: ./media/version-compat-tests/runtest.sh 1
Change-Id: I390ebb0d9bc512c978b4ae3c57b0e7e0d8c1b68c