Update patch set 4
Patch Set 4:

> > A) Sergey could tell us more about the advantages of having all of the singletons on a RootProjectExtension

As Jake is against global state, I'm against of getTaskByName (as you see we're against of different things, that don't contradict each other). So, why I'm against of getTaskByName?
once you use getTaskByName all over the place, any code can query task by name and do something with it without thinking about an layering and abstractions etc (because everything is always public, why not grab some task in the middle of pipeline and do something with it?). Avoiding taskByName makes you think what you want to expose and what you want to hide and what you actually should use. 

> I don't know if we'll be able to reconcile the extension on the root project that knows about every task with Gradle's lazy task functionality we'll want to migrate towards. But again, I don't think that's a strong argument against it since we can do another migration.


I'm not sure if could be the one extension or multiple different extensions or, but I'm positive that it is possible to do in more typesafe and explicit way and still have lazy task functionality: we can expose TaskProviders instead, it is close to what android gradle plugin does.

Patch-set: 4