Update patch set 5
Patch Set 5: Presubmit-Ready+1

I discussed this change with Brian Carlstrom from ART team. He said it's okay to do so if we are going to unhide this methods in Q. As otherwise it assumes it will work just before name and signature of the hidden method will never be changed which is not guaranteed for hidden methods.
Yigit, could you please +2 if you approve this changes?
George, FYI we should not forget to unhide the getter(isLayoutSuppressed) as well, not only the setter(suppressLayout) which is used in Transitions currently.

Patch-set: 5
Reviewer: Gerrit User 1361834 <1361834@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1