Add a composable interface for ComponentActivity#onBackPressed
Allow any class with a reference to a ComponentActivity
to add/remove an OnBackPressedHandler interface for
intercepting and handling the onBackPressed callback in
the ComponentActivity.

Each add call also passes in a LifecycleOwner, ensuring that
the handlers are automatically cleaned up if their
associated Lifecycle is destroyed. This prevents memory leaks
and makes it appropriate for use in child lifecycles such as
those of a Fragment.

Test: new ComponentActivityOnBackPressedTest tests
BUG: 117990885
Change-Id: I2b283b8a084b5ddec3801d645afb2c6bc5c02657