Update patch set 2
Patch Set 2: Presubmit-Ready+1

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > we should have a test so that this does not happen again.
> 
> maybe you can test by swapping the main thread arch executor w/ a dummy one that throws executes and run a query from main thread (using runOnUI) and ensure it runs?

Wrote a test!

Was able to write a decently real test that uses StrictMode to detect disk IO in the main thread. Sadly a good way of detecting the violation is only available in API 28, hence the suppress SDK.

Moreover this took a while to do because exception thrown by beginTransaction() in the InvalidationTracker where being aborted because the invocation was inside the try-finally used to invoke endTransaction(). Specifically if beginTransaction() threw an exception such a StrictModeViolation or some other exception (maybe a SQLiteFullEx) it would get aborted because then endTransaction() in the finally block would also throw an SQLiteException along the lines of 'no transaction is active'.

Patch-set: 2
Reviewer: Gerrit User 1258207 <1258207@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1