leanback: do not setup guidedstepsupportfragment transition in constructor
onProvideFragmentTransitions() is already called in onCreate().
Calling in constructor is redundant. Remove the call in constructor
to better work with lifecycle aware component which is created in
onCreate().

However for deprecated framework fragment, setup fragment
transition in onCreate() does not work in most of verions, here is
a list of test result:
- API 23: enter transition set before onCreate() will be honored
- API 25: enter transition set in onCreate() will be honored
- API 27: enter transition does not work
- API 28: enter transition does not work
Given it's too complicated and deprecated, we do not make changes.

Test: launching GuidedStepSupportFragment from BrowseSupportFragment
test of SupportLeanbackDemos and transition is still working.
Bug: 117347911

Change-Id: Ie12c122c419475da26d851ed3849bc199eda49c0