Create CachedSliceLiveData.
CachedSliceLiveData is a new public class that extends the previously
used LiveData<Slice> so that more control over the caching process is
given to the presenting app.

Currently it allows the presenting app
to ensure the input stream has been loaded, and allows the caller to
force the app to switch to a live state regardless of whether anything
has been clicked.

Change-Id: I58d482cc6de552c666ec55155e02a8495c55bb22
Fixes: 111893419
Fixes: 112697672
Test: Existing + new tests in CachedSliceLiveDataTest.