Update patch set 42
Patch Set 42:

Presubmit build is failing with this error:

Execution failed for task ':textclassifier:integration-tests:testapp:preMaxDepVersionsDebugAndroidTestBuild'.
> Conflict with dependency 'com.google.code.findbugs:jsr305' in project ':textclassifier:integration-tests:testapp'. Resolved versions for app (1.3.9) and test app (2.0.1) differ. See https://d.android.com/r/tools/test-apk-dependency-conflicts.html for details.

I am not sure if it is related to the change I made in core/build.gradle that is causing a chain of dependency conflicts, or it is a separate issue. Any ideas?

Patch-set: 42
Reviewer: Gerrit User 1140244 <1140244@85c56323-6fa9-3386-8a01-6480fb634889>