Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Could we test this to make sure we don't regress?

The existing tests are sufficient, as long as they are run on the SDK versions that contain the bug (this is how this bug was reported). Beyond that, I can't think of a reasonable way to reproduce the problem. Any ideas?

Patch-set: 3
Reviewer: Gerrit User 1061011 <1061011@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0