Update patch set 1
Patch Set 1:

> 1) Could you cherry pick the Metalava change into this branch, generate the files, and copy them back to master? Or, alternatively, what's the plan for "manually" generating them since master has already moved on?

I had a checkout before master moved on, so that CL is here: https://android-review.googlesource.com/c/platform/frameworks/support/+/805436

In general, when we don't have metalava changes, I think there's a process question - who owns that cherry-pick-the-api-files part of the release? In the file-a-bug, that makes sense for the bug to encompass that, but otherwise, is it just bookkeeping everyone with a release branch should be aware of?

> 
> 2) Follow the file-a-bug process. Since this is a special case, we can run it outside of the monthly release schedule and hit the two-week cadence directly. If you already qualify for RC (e.g. API surface reviewed, zeroed out P0/P1 bugs, 100% certain you won't add new APIs) then we could technically skip the beta cycle. If you don't absolutely have a business need to release this RIGHT NOW, though, we should take it easy and err on the safe side with a slow and complete release process. Again -- special case. In every other case we'd do the normal monthly (eventually bi-weekly) release cycles.

We have zero bugs filed on this release (it's pretty small), no new APIs planned.

I'd really like to hit the beta or RC asap to encourage use and shake out anything remaining before I go OOO in two weeks. Will file the bug for beta.

Patch-set: 1
Reviewer: Gerrit User 1010334 <1010334@85c56323-6fa9-3386-8a01-6480fb634889>