Fix fontFamily not working on pre API 24
For API 24 and before applying a fontFamily programatically would not work
correctly. If we had an xml that described using a font for weight 400
and actually a different one for weight 700 (bold) then in reality we
would retrieve the first one and when needing the bold variant we
would just bold it instead of using the second one.

For API 21-24, used reflection to call addFontWeightStyle.
For pre API 21, mapped Typefaces to font family entries at
deserialization and retrieved them in AppCompatTextView#SetTypeface.

Bug: 71565273
Test: ./gradlew core:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=androidx.core.graphics.TypefaceCompatTest\#testApplyFontFromFamilyUsingStyle
Change-Id: I2996b36225190e86300a2df6f62450151ef58f8a