Make ComponentActivity#getLifecycle() final
Much of ComponentActivity (as well as classes
built on top of it) assume there is a single
Lifecycle object for a given Activity.

This object is controlled by the ComponentActivity
and any attempts to introspect or override it
invariably is the wrong thing to do. Therefore
by making this final we avoid subclasses doing
the wrong thing and can build upon the
guarantees that having a final instance provides.

Test: ./gradlew bOS
BUG: 78542858
BUG: 118690391
Change-Id: I11768cd4de1c845d9ddc92dd8b451a01b9bd8e5e