Update patch set 2
Patch Set 2: Code-Review-1

I think this is an example of something that should be fixed in RV.  RV is sending the events, and it assumedly knows that it sent a settling event, and if it fails to follow that up with an Idle event.. well then that is RV's fault.  I think your fix probably fools RV into doing the right thing, when the better fix would be to have RV do the right thing in the first place. Right?

Patch-set: 2
Reviewer: Gerrit User 1344416 <1344416@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1