Update patch set 1
Patch Set 1:

I think we should make all the elements toggle themselves off when there's not enough space. It'd fine to group the up/down buttons together.

But the next step should be measuring the thumb. If there's not enough space for that, then don't draw it, but we can still have the up/down buttons.

And same for the alpha jump button.

Patch-set: 1
Reviewer: Gerrit User 1344512 <1344512@85c56323-6fa9-3386-8a01-6480fb634889>