Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> I think this is an example of something that should be fixed in RV.  RV is sending the events, and it assumedly knows that it sent a settling event, and if it fails to follow that up with an Idle event.. well then that is RV's fault.  I think your fix probably fools RV into doing the right thing, when the better fix would be to have RV do the right thing in the first place. Right?

Sounds right, indeed. There is one property though that only occurs with a SnapHelper: even though there is a high velocity when releasing your finger after a fling, the RV must stop immediately. In general, is it still the responsibility of RV to deal with cases that can only occur in the presence of an external influencing components? And how will you test it, if it is not possible to trigger the behavior without another component?

If you feel this should be part of RV, and you know how to test it without PagerSnapHelper, please go ahead. In that case, I'll add SwipeToLocation to testutils in a different CL where I also need it.

Patch-set: 2
Reviewer: Gerrit User 1347655 <1347655@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0