WIP Fix live data observe forever in room
This CL fixes a problem where if DAO returned LiveData is
observed with observeForever but there are no references
to the LiveData, it would eventually be garbage collected.

To avoid this, InvalidationTracker now manages these LiveData
objects and strongly references them as long as there is an
active observer.
To be 100%, we should actually do it when LiveData has observers
but right now that API does not exist in the LiveData side
to get notified when the observer count changes 0 and 1.

I've also cleaned up some test code to avoid duplicating the
test observers in each test.

I also needed to move more dependencies into prebuilts to
fix the min-max version check issues. We should've moved to
prebuilts anyways.

Bug: 74477406
Test: LiveDataQueryTest, InvalidationTrackerTest
Change-Id: If88114306d7e20ddaed0312a1203a0d14b710145