XMediaPlayer: Fix flaky AudioFocusTest
The noisy intent receiver, which pauses or ducks the volume with the
noisy intent, has been registered when the playback state becomes
PLAYER_STATE_PLAYING. Because of that, followings may happen.

  Step 1) XMediaPlayer#play() is called.
  Step 2) Noisy intent is broadcasted --> Ignored because receiver
          hasn't registered.
  Step 3) XMediaPlayer starts playing with the normal volume.
          XMediaPlayer.PlayerCallback#onPlaybackStateChanged(PLAYING)
          will be called.
  Step 4) Register noisy intent receiver in the callback. But it's after
          the noisy intent.

This CL fixes the issues by handling noisy intent receiver and audio
focus when play/pause/reset are called, instead of handling when the
player state is changed.

Note that this matches more with the media session callback guideline
published in the developer's site.

Bug: 117352673
Test: $ ./gradlew media2:connectedCheck (5 times)
Change-Id: Icbb6e1481886cb04a2a24c482cb6af1e639b5190