Update patch set 52
Patch Set 52:

So I went ahead a fix some bits in this CL to unblock you. 
So things I fixed and some drive-bys as a result:
- got read of the lock in ShorcutsContainer. It was dictated by correctness (not philosophical point of view). You had a lock and mutated a field (mIcon) on ShotcutInfoCompat with this lock, however, saver didn't control that ShotcutInfoCompat object, so that someone can easily access that field without taking a lock. It could be fixed by defensive copy, but since I started fixing, I got read of it mutation with concurrency, just because it is error-prone
- return ListenableFuture instead of ResolvedFuture. (Map / HashMap situation)
- removed useless synchronized keywords
- moved serialisation / deserialisation logic to separate file;

Things that would be nice to fix going forward, but not in this CL:
- cancelation (operations ignore cancelation signals =\, but it will be easier to fix with CallbackToFutureAdapter)
- initialisation (technically we lose exception in some situations there, but pragmatically it is ok)

Another thing,
do you want these files to be backed up or not? As far as I remember by default everything in getFilesDir() is backed up, not sure what is the expected behavior.

Patch-set: 52
Reviewer: Gerrit User 1098853 <1098853@85c56323-6fa9-3386-8a01-6480fb634889>