Update patch set 3
Patch Set 3: Code-Review+2

overall implementing the feature in framework but disabling it with support library and falling back to staticlayout is a little weird unfortunately. 

I think we should try to get a fix to AOSP/Framework P asap. So that the phone manufacturers can use the correct version. 

Moreover if we can make the fix contained, we can ask release as an update. This is something that deserves to be escalated. 

Nona, can you please check what we can do for those items above?

Patch-set: 3
Reviewer: Gerrit User 1106808 <1106808@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2