Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> What is this fixing? I'm not sure we actually want this, since a view can only move in whole pixels

It's not so easy to describe but have a look here, there are gifs with actual and proposed behavior: https://bitbucket.org/bio007/coordinatorlayout_rounding_bug

The point is that if you animate some view (with translation for example) and the animation step is small (with decelerator interpolator for example) you won't get onDependentViewChanged method called in behavior. We have this problem with bottom sheets. If we depend on bottomsheet view and want to change something when the bottomsheet is hidden -> moved to 0 position, we don't get correct notification because last change is 0.9 or 1.6 or some random number which is close to 0. The movement between that last point eg. 0.9 and final value 0 is so small that CoordinatorLayout ignores it at all.

Patch-set: 1
Reviewer: Gerrit User 1197023 <1197023@85c56323-6fa9-3386-8a01-6480fb634889>