Update patch set 1
Patch Set 1:

Hmm, after getting this working, I realized that Metalava has a setting for generating a private API but I don't see one for generating only the portion of the API that is public but marked with RestrictTo(LIBRARY_GROUP).

I assume that it's important to do the latter (instead of the former, which is what this change does). Right?

Patch-set: 1
Reviewer: Gerrit User 1137668 <1137668@85c56323-6fa9-3386-8a01-6480fb634889>