Fix event sequence on config change during scroll
Fixes the sequence of events that are fired by ViewPager2 when a
configuration change happens during a smooth scroll to a page.

Required fixes in the tests as well: tests can now supply a callback to
recreateActivity so they can configure the new ViewPager2 instance
during the TestActivity's onCreate. This is used to reattach the
RecordingListener directly after a config change.

Bug: 116661075
Test: ./gradlew viewpager2:connectedMaxDepVersionsDebugAndroidTest
Change-Id: Ia7c0c3fb26db444f9454c38ba1ee14522a167252