Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Could you give me a quick explanation as to why this is needed?  I'm sure it's fine, but I want to understand before +2.

This is part of managing library-specific releases.

If kept as their dynamic project versions, recyclerview would depend on Tip-Of-Tree versions of both libraries.  Given that we are not releasing annotations-1.1.0-alpha01 or customview-1.1.0-alpha01, we would be depending on libraries that are not being released.

We ARE releasing core-1.1.0-alpha02, that is fine to be left in dynamic project version.

Patch-set: 1
Reviewer: Gerrit User 1364182 <1364182@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0