Update patch set 6
Patch Set 6:

Technically, the fact that this change automatically enables running checkApiRelease against restricted APIs is slightly different than we discussed, although it adds a flag "checkReleasedRestrictedAPIs = false" that can be added into a build.gradle to turn it off on a per-project basis.

My thought is that if we start generating the api/system*.txt files now then we can see whether this workflow (of having to regenerate the api/system*.txt files) is inconvenient but if it becomes a pain that we haven't yet auto-migrated everything (from RestrictTo(LIBRARY_GROUP) to RestrictTo(LIBRARY)) then I'm supportive of setting "checkReleasedRestrictedAPIs = false" in other projects too (or changing the default if it's super inconvenient).

Patch-set: 6
Reviewer: Gerrit User 1137668 <1137668@85c56323-6fa9-3386-8a01-6480fb634889>
Work-in-progress: false