Deprecate RVAXDelegate and Make sure adding accessibility actions to recyclerView Items works
Nearly every client of RecyclerViewAccessibilitYDelegate is doing something we don't want them to do in API (tracking focus, changing row count semantics),
work around bugs that existed at one point or another in frameworks,
or doing something that is much more easily done in newer API (eg addAccessibilityAction).
Getting rid of this needlessly complex API reduces the ways a developer can shoot themselves in the foot, and makes it eaiyer for them to make their recyclerViews accessible.

For example, if the user used the old AXDelegate API, and modified the item delegate, they would need to keep track of layoutManager state to make sure a crash doesn't happen on animation.
With the new API they follow the same pattern as adding an onClickHandler and they avoid that class of bug completely.

I added tests to demonstrate adding and replacing is functional and easy.

Test: See the unit tests.
Bug: 31493845
Change-Id: I13446b02d52653600c02bc6af19294840e281b03