Media: Use PID information from the controller
Binder#getCallingPid() always returns 0 for oneway calls from a remote
process, and we can only get when the controller has sent the PID.

Due to the change, PID may not be available (e.g. MediaBrowserCompat
compiled with the older supportlib library), so we shouldn't use
Framework's RemoteUserInfo#equals() directly for comparing two objects.

Bug: 111418518, Bug: 111418519
Test: $ ./gradlew media:connectedCheck media2:connectedCheck
      $ media/version-compat-tests/runtest.sh 4
      $ media2/version-compat-tests/runtest.sh 1
Change-Id: I8733a4d8538bcb482c177240bdbd60e6625c3d5d