Fixed testing with flavors
Projects now add themselves manually to a module name to project
name mapping so we don't have to guess module from the project
name.

All configurations with "maxDepVersions" in their names will use
tip of tree dependencies instead of androidx artifacts.

Removed @aar from androidx dependency names from Dependencies.kt
as a workaround fix for a Gradle bug which happens sometimes
when aars are substituted with project versions of the library.
MavenUploadHelper will now know if an androidx external
dependency is of type aar and if so will include it in the POM.

Test: Introduce a bug in tip of tree version of a library and
ensure that running minDepVersion tasks are not affect by
that bug's existenece and ensure the opposite for maxDepVersion's
tasks.

Bug: 112531649, 118385540
Change-Id: Ic76bc89e8f9637afeee994f0827c650926377e7e