Ensure retained Fragments are properly removed
Avoid an IllegalStateException caused by
retained Fragments not being properly removed
from the FragmentManager.

This only occurs when using a FragmentHostCallback
that implements ViewModelStoreOwner (which is
the case for FragmentActivity), so switched
FragmentLifecycleTest over to using that model
to reproduce and confirm the bug fix.

Since FragmentLifecycleTest does not actually
create and destroy real FragmentActivity instances
but instead relies on the FragmentController API,
moved FragmentManagerImpl from using a hard coded
Activity.isChangingConfigurations() dependency to
passing that value in explicitly in
FragmentController.dispatchDestroy(). Besides
testing, this has the added benefit of supporting
non-FragmentActivity uses cases such as Android
Auto.

Test: updated FragmentLifecycleTest
BUG: 119592051
Change-Id: I6a543b53850df5560f45f1644193df0c465273f9