Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I have a bold suggestion, so tell me what do you think about it:
> 
> - Move body of mOnBackPressListener#onPopBackStack in NavController to NavController.syncBackStackToExternalModification(Navigator), that can be public or hidden.
> - Add backpressed listener in NavHostFragment and call syncBackStackToExternalModification from that listener.
> 
> It looks like it should work and we get rid of that special hidden infrastructure and code gets way easier. The downside I see it is harder to hook up FragmentNavigator

This whole API is going away once NavController can rely on OnBackPressedCallback as this approach is flawed in general (see b/111598096), but is the best we have for 1.0. I'd rather not invest in another approach at this point.

Patch-set: 1
Reviewer: Gerrit User 1060451 <1060451@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0