Remove Context from most Navigators
NavGraphNavigator only took a Context for
providing a human readable resource name in
debugging calls. Since the primary use case
for using resource ids is by inflating
navigation graphs from XML, we take advantage
of the existence of a Context in onInflate()
to get a human readable name, avoiding the
need to provide a Context into NavGraphNavigator
and classes that rely on it, such as
TestNavigationProvider.

FragmentNavigator also get its Context parameter
removed as it was unused.

Test: updated tests still pass
Change-Id: I0270c4eaf2fedb4f86753402008b338c14c2ef01