Update patch set 6
Patch Set 6:

> Patch Set 6: Code-Review-1
> 
> can we add these after the lint check works instead of before ?
> 
> I'm a bit bitter on the time i spent (aosp/826955) because of the min-max API thing which does not yet work but breaks presubmit-project dependencies for tests. 
> 
> It would be really nice not to merge OR enable these until they are functional to avoid the extra inconvenience without the benefits.

The issue with aosp/826955 has nothing to do with the "min-max API thing". It's a Gradle issue that would have happened as soon as we started migrating to artifacts instead of prebuilts with or without the build flavors etc... They are absolutely unrelated, and although min-max testing wasn't functioning properly before, it causes no problems what so ever. I think I should have made it clear a while ago but I assumed everyone was aware of what was happening. Given that the perception was that the min-max testing was causing this, it makes sense why the change in aosp/825804 was made inside the useMaximumDependencyVersions function. That problem was also one we faced and spent almost two weeks trying to address, we are also bitter about it.

Patch-set: 6
Reviewer: Gerrit User 1354330 <1354330@85c56323-6fa9-3386-8a01-6480fb634889>