Update patch set 8
Patch Set 8: Code-Review+2

(3 comments)

As Clara mentioned, afaik compound drawables was a remedy for an older issue/view hierarchy. We do not want to invest more in them. Since Nick added compat support for an existing TextView API I think it is not in the same category of improving the compound drawables. More importantly the CL is already prepared, therefore I think it is good to have this support.

Patch-set: 8
Reviewer: Gerrit User 1106808 <1106808@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2