Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> lets document how ignoredColumns work for embedded.
> also, maybe, decide on it ? :)

Added documentation to ignoredColumns.

To be consistent ignoredColumns should work as close as @Ignore as possible. With @Ignore you can't ignore individual @Embedded columns, similarly the user won't be able to use the name of an embedded POJO column to ignore it, instead the user has to use the name of the field which will then ignore all of the columns the embedded field would have added. This is similar to placing @Ignore in the @Embedded POJO field. Moreover, ignoring a single column of an embedded field breaks the possibility of constructing that POJO, so its not very reasonable to allow such thing.

Patch-set: 1
Reviewer: Gerrit User 1258207 <1258207@85c56323-6fa9-3386-8a01-6480fb634889>