Update patch set 3
Patch Set 3: API-Review+1 Code-Review+2

There's a decent chance that this will catch on fire when we attempt to push it to google3, but it is binary compatible. Just be prepared for me to come back with a P1 bug that requires you do a large-scale change in google3 (not as bad as it sounds).

Patch-set: 3
Reviewer: Gerrit User 1044446 <1044446@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2
Label: API-Review=+1