[WebView Support Library] clearer callback timeout messages
This adds a helper to create clearer callback timeouts, and centralizes
logic for timeouts.

This also documents that the helper should only be used with @LargeTest
(which we currently violate), although our bots and local scripts seem
to run long tests to completion regardless of @{Small,Medium,Large}Test
annotation, so this doesn't change existing uses.

Bug: 120167097
Test: ./gradlew :webkit:connectedAndroidTest
Change-Id: I380bd8ce0879b11ed90c7c12eeb1a4f40b6c8a7f