Update patch set 2
Patch Set 2:

I think it's reasonable to centralise like this, but I wonder whether a better option might be to migrate to futures as a means of expressing these flows? I'm not sure if you can write tests with futures in CTS, though.

Is it really an @LargeTest if only the failure case is slow, given that we expect tests not to fail?

Patch-set: 2
Reviewer: Gerrit User 1101490 <1101490@85c56323-6fa9-3386-8a01-6480fb634889>