Create a new Lifecycle for Fragments after destruction
After a Fragment is destroyed (i.e., due to a
configuration change or removal from the FragmentManager),
the existing Lifecycle should remain in the terminal
DESTROYED state.

However, Fragments can be reused basically as a completely
new instance. In the initState() method which resets the
Fragment's state, we should also reset the Lifecycle
and SavedStateRegistry to ensure that the 'new' instance
is entirely independent from the old instance.

This fixes issues where developers would reuse a Fragment
and lifecycle.currentState would be DESTROYED in onCreate(),
causing any LiveData.observe calls to be ignored.

Test: updated FragmentLifecycleTest
BUG: 118880674
Change-Id: I7653eade8bac1a1101f3bba96198aae6a32ee264