Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Originally idea was that registered lifecycle observers follow all lifecycle events for the given instance, so if you reuse the instance, observers still work and you can add your observer in construction time and never think about it again. Maybe we can simply mark our state INITIALIZED in initState?

Given the lengths we went through for the Fragment view lifecycle to ensure that DESTROYED is a terminal state for that specific Lifecycle object, resetting the Lifecycle back to INITIALIZED seems like the wrong thing to do.

Are you suggesting that having DESTROYED as a terminal state is something we want to remove as an invariant on Lifecycle?

Patch-set: 1
Reviewer: Gerrit User 1060451 <1060451@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0