Update patch set 1
Patch Set 1: Code-Review+2

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > So I'm not sure what this will really tell us other than a wakelock that we think shouldn't be held is still held.  I don't think it actually gives you any diagnostics.  Why should we check this in?
> 
> I thought this log message in the context of other logs will tell us more. Did you have more diagnostics in mind ?

I just wanted to be clear there wasn't anything more in terms of diagnostics that was incoming.  Since we've gotten no real bugreports with this, this is really only going to be for us (and we haven't necessarily observed this phenomenon).

Patch-set: 1
Reviewer: Gerrit User 1060206 <1060206@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2