Fix LifecycleCallback leak and PerformException crash
The ActivityTestRule wasn't properly finished at the end of tests,
resulting in its internal LifecycleCallback not being removed from the
ActivityLifecycleMonitorRegistry. It is now an annotated rule property,
so Espresso will manage the rule for us.

When an exception occurs during swipe injection, it is wrapped in a
PerformException. The builder needs at least an action and view
description, or it will crash when building the PerformException.

Bug: 120472989, 120473551
Test: ./gradlew viewpager2:connectedMaxDepVersionsDebugAndroidTest
Change-Id: If89d23c0a6a6da5b8f4e8a5b44eaf11d832cad12