[WebView Support Library] Add test for onReceivedHttpError
This adds a test for onReceivedHttpError. This test was originally
omitted because we didn't have a web server available for AndroidX
tests. This blocker was fixed by laisminchillo@ in http://aosp/795653.

This adds okhttp3's MockWebServer library to AndroidX dependencies
(copied from laisminchillo@'s other CL: http://aosp/795891), and
reimplements the CTS test.

Bug: 119592680
Test: ./gradlew :webkit:connectedAndroidTest
Change-Id: I1a666f3b570dfc30d30d4f28cbcac2f992290a58