Update patch set 3
Patch Set 3: Code-Review-2

This is not the cause for the failure. If you look at the screenshot in the failures, it is failing due to r.android.com/838308 as it fails with "androidx.activity.integration... | This app was built for an older version of Android ...". We need to add <uses-sdk android:targetSdkVersion="${target-sdk-version}"/> to AndroidManifest of that project.

Patch-set: 3
Reviewer: Gerrit User 1086749 <1086749@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2