Fix throwing Exception in getCurrentBrowserInfo/getBrowserRootHints
When framework browser is connected to MediaBrowserServiceCompat,
calling getCurrentBrowserInfo()/getBrowserRootHints() in service
callback methods throws IllegalStateException, which results the
crash of the service.

This is because the current connection info is only set when compat
browser is connected, not for framework browser.

This CL sets a dummy connection info right before the service callback
is called by the framework browser, so that calling those methods
does not throw an exception.

Bug: 120638640
Test: media/version-compat-tests/runtest.sh 4 (on API 23/24/26/28)
      media2/version-compat-tests/runtest.sh 1 (on API 24/28)
Change-Id: I1f011d68f4e40aff2f87bfee661f6fd71e0ee227