Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> What is this for? Please give me more context.
> BTW, I'm hesitate for adding jungshik@ to our Android duty.

triage-guesser.py is a not official but slightly convenient script for guessing the assignee of a bug based on the title of the bug, without having to reread the triage doc when it changes.

Clara mentioned that she updated the assignees in the triage doc, so I looked and found a line that says "- Fonts: nona, dougfelt (i18n MTV), jungshik (i18n MTV)".


I'll remove junkshik for patchset #2

Patch-set: 1
Reviewer: Gerrit User 1137668 <1137668@85c56323-6fa9-3386-8a01-6480fb634889>