Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1
> 
> (1 comment)

This is currently inactive (not running as part of errorprone) and we are not yet sure what policy we will be employing for SDK_INT comparison. @alanv and I talked about this this morning and decided to wait for @jakew to be back to discuss this since there are different opinions regarding what the rule should be. Either way the rule will be ported to lint, so there is no reason for this one to stay.

Patch-set: 2
Reviewer: Gerrit User 1354330 <1354330@85c56323-6fa9-3386-8a01-6480fb634889>