Update patch set 3
Patch Set 3:

(1 comment)

I tried the demo with and without this CL and it didn't behave like a ViewPager. When I changed the class name to the old package (android.support.v4), it started working. I guess I have an older version of TalkBack, then? How do we deal with backwards compatibility?

Patch-set: 3