Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> Not sure I understand the context. I don't see when the InvocationHandler would actually be null (this is provided by chromium [1], right?)
> 
> [1] https://cs.chromium.org/chromium/src/android_webview/support_library/java/src/org/chromium/support_lib_glue/SupportLibWebViewChromium.java?l=80&rcl=05fe2faf641bb9c29fe9c454c8e3e6f0f7a65e80

You're right. The way this works is slightly opaque; the invocation handler is always non null, but the adapter that it is an invocation handler for might reference a null peer, and in that case getOrCreatePeer returns null on the remote side rather than constructing an object that wraps the invocation handler.

Definitely forInvocationHandler should be @Nullable (returns null in single process) but we don't need to handle the invocation handler being null.

Patch-set: 2
Reviewer: Gerrit User 1101490 <1101490@85c56323-6fa9-3386-8a01-6480fb634889>