Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I'm trying to write a test that actually uses assembleDebug task so that we don't regress but is turning out to be quite hard. Stay tuned!

Wouldn't testing that generateDebugSources runs the task be enough? That'll ensure that the plugin is being called (as we know that generateDebugSources is run by assembleDebug - no need to test that) and if the plugin succeeds, we know that the ApplicationIdWriterTask will have run as well.

Patch-set: 3
Reviewer: Gerrit User 1060451 <1060451@85c56323-6fa9-3386-8a01-6480fb634889>