Update patch set 7
Patch Set 7:

While addressing Insun's comment and handling the FD life cycle in MediaPlayer w.r.t. playlist, I realized that taking over the ownership of FDs from the client might not be a good idea since there some cases the player cannot know the clients intention.

e.g. when there is a playlist of several songs and a user changes position of the song from A to B, the client would call "item = removePlaylistItem(A);" and "addPlaylistItem(b, item);", but in this case, MediaPlayer has no idea about whether the item will be readded or not, so, MediaPlayer would close the underling FD and we would need to require developer to create new FileMediaItem/FD in this case.

This seems to be very error-prone and hard to enforce. So, the current thinking is to just let developer/client keep the ownership of the FD passed to MediaPlayer. Let me know if you have concern or different ideas on the plan.

Patch-set: 7
Reviewer: Gerrit User 1026966 <1026966@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0