Update patch set 1
Patch Set 1:

I'm still not convinced that mocking is the right answer. While it is not harmful in simple case like you suggest in the video, once developers add mockito to the project, they are tend to cross the line and mock things that shouldn't be mocked and have complicated behavior, so in result their mocks don't have much common with reality and tests have low value. So simply advocating for "no-mockito" would be much easier story.

Patch-set: 1
Reviewer: Gerrit User 1098853 <1098853@85c56323-6fa9-3386-8a01-6480fb634889>