Media2-Widget: Fix media2.VV.extractMetadata()
This CL fixes a crash when VV.setMediaItem() is called multiple times in
a short time.
The crash happened when onInfo() is called for the previous item but
mediaplayer is not iniaizlied due to new setMediaItem() calls.
(mMediaPlayer.getTrackInfo() throws exception in this case.)

To prevent that, we check callback argument and mMediaItem.

There was also an issue that mMediaItem can be changed during extracting
metadata.
Adding an argument to extractMetadata() to fix that.

Bug: 120946772
Test: ./gradlew media2-widget:connectedCheck
      && ./gradlew media2-widget:lint

Change-Id: I090e1b5a44b6b6b0bebc8da38ba8fc82058d09e9