Safe Args in Kotlin
This change allows users to apply a version of the Safe Args Gradle
Plugin that will generate code in Kotlin. The id of the plugin is
'androidx.navigation.safeargs.kotlin'.

The API for the Kotlin version of the Directions and Args classes is
different than the Java version. The main difference being the lack
of builders and setter. Args and NavDirections classes are immutable
data classes and default values are default parameter values with the
idea that invocations are done with named parameters.

Bug: 110263087
Test: ./gradlew navigation:navigation-safe-args-gradle-plugin:test \
                navigation:navigation-safe-args-generator:test

Change-Id: Ie5f7f63c52dc0630cfe8bfe4fddcbe1181bc22d4