Update patch set 27
Patch Set 27:

(9 comments)

Hey Michal,

I know it is a lot of work to make a contribution to RV like you are.  If you'd rather pass the responsibility of writing adequate tests (hugely important given the complexity of RV) onto me at this point.  That is fine.  To do so, what I'd ask is that you mark all of the new public api methods as "@RestrictTo(LIBRARY_GROUP)" and all docs as "@hide", OR make them all private and remove usage from the car library for now (the latter is a bit safer, otherwise Car library is using code that isn't fully tested).

Then we can just submit the code as is, and I can follow up with a new CL that adds the tests I think should exist for it to be public API.

Let me know.

Patch-set: 27
Reviewer: Gerrit User 1344416 <1344416@85c56323-6fa9-3386-8a01-6480fb634889>